Update README.md
This Pull Request introduces a comprehensive update to the README file of the SmartFramework repository. The changes include an enhanced project description, detailed feature list, clear setup and installation guidelines, contribution procedures, and licensing information. This update aims to provide a more informative and user-friendly introduction to SmartFramework, making it easier for new users to understand the project's purpose, use it effectively, and contribute to its development.

Issue: #8